refactor: fastapi 관련 구조 refactor
