+++ src/main/java/com/teamEWSN/gitdeun/common/fastapi/dto/FastApiCommitTimeResponse.java
@@ -1,4 +1,4 @@
-package com.teamEWSN.gitdeun.common.config.fastapi;
+package com.teamEWSN.gitdeun.common.fastapi.dto;
 
 import lombok.Getter;
 import lombok.Setter;
@@ -7,6 +7,6 @@ import java.time.LocalDateTime;
 
 @Getter
 @Setter // JSON 역직렬화를 위해 필요
-class FastApiCommitTimeResponse {
+public class FastApiCommitTimeResponse {
     private LocalDateTime latestCommitAt;
 }
